pyproject.toml
[project]
name = "swarmy"
version = "0.1.0"
description = "Swarm robotics simulation"
requires-python = ">=3.13"
dependencies = [
    "pygame",
    "pyyaml",
    "numpy",
    "matplotlib",
    "scipy",
]

[dependency-groups]
dev = [
    "ruff",
]
